+++ jsolex/src/main/java/me/champeau/a4j/jsolex/app/script/JSolExScriptExecutor.java
@@ -25,6 +25,7 @@ import me.champeau.a4j.jsolex.processing.expr.DefaultImageScriptExecutor;
 import me.champeau.a4j.jsolex.processing.expr.impl.FileSelector;
 import me.champeau.a4j.jsolex.processing.expr.impl.Loader;
 import me.champeau.a4j.jsolex.processing.sun.Broadcaster;
+import me.champeau.a4j.jsolex.processing.sun.workflow.PixelShift;
 import me.champeau.a4j.jsolex.processing.util.ImageWrapper;
 
 import java.io.File;
@@ -37,7 +38,7 @@ import java.util.function.Function;
  * A script executor which exposes a JavaFX file chooser to the {@link Loader}.
  */
 public class JSolExScriptExecutor extends DefaultImageScriptExecutor {
-    public JSolExScriptExecutor(Function<Double, ImageWrapper> imageSupplier,
+    public JSolExScriptExecutor(Function<PixelShift, ImageWrapper> imageSupplier,
                                 Map<Class, Object> context,
                                 Broadcaster broadcaster,
                                 Stage stage) {
@@ -45,7 +46,7 @@ public class JSolExScriptExecutor extends DefaultImageScriptExecutor {
         prepareContext(context, stage);
     }
 
-    public JSolExScriptExecutor(Function<Double, ImageWrapper> imagesByShift,
+    public JSolExScriptExecutor(Function<PixelShift, ImageWrapper> imagesByShift,
                                 Map<Class, Object> context,
                                 Stage stage) {
         super(imagesByShift, context);